requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
anthropic==0.79.0
httpx==0.27.2